Improve Ethereum client wrappers adoption
Changed the way Ethereum client is used
in the Ethereum chain implementation.
Specifically, got rid of the bare client
usages. Instead of that, the wrapped client
which provides logging and rate limiting
capabilities is used everywhere.
