+++ go.mod
@@ -18,7 +18,7 @@ require (
 	github.com/gogo/protobuf v1.3.1
 	github.com/google/gofuzz v1.1.0
 	github.com/ipfs/go-log v1.0.4
-	github.com/keep-network/keep-common v1.2.1-0.20201020114759-19c123cbd4f4
+	github.com/keep-network/keep-common v1.2.1-0.20201103151750-c52f5d480c10
 	github.com/keep-network/keep-core v1.3.0
 	github.com/keep-network/tbtc v1.1.1-0.20201026093513-cb9246987718
 	github.com/pkg/errors v0.9.1
